App now works with internal build
